requirements.txt
fastapi
uvicorn
python-docx
python-multipart
openai
google-genai

python-dotenv
sqlalchemy
PyJWT
passlib[bcrypt]
requests
aiofiles
jinja2
